% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cli}
\alias{cli}
\title{cli data}
\format{
A matrix:
\describe{
...
}
}
\source{
\url{NULL}
}
\description{
The cluster information of BLCA sample from TCGA, GSE32894, GSE13507, and E-MTAB-4321
}
\details{
Success!!
}
