% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test}
\alias{test}
\title{test data from GSE32894 for test}
\format{
A matrix:
\describe{
...
}
}
\source{
\url{NULL}
}
\description{
a array data from GSE32894 for test
}
\details{
Success!!
}
